refactor(audio): split into backend / playback / procedural + position/orientation overloads
Two changes in one commit:

1. **Module split.** `audio/audio.ts` was a ~1100-line file mixing
   everything from file loading to procedural synths. Split into four
   files along functional seams:

   - `audio/backend.ts` — shared internal state (the `state` object
     replacing the scattered module-level `let`s), `soundLoadError`
     helper, `stopOnAudioError` flag, and the two public WebAudio
     escape hatches (`getAudioContext`, `getMasterGain`).
   - `audio/playback.ts` — file-based playback: `load`, `play`, `fade`,
     `seek`, `rate`, `stereo`, `position`, `orientation`, `panner`,
     `stop`, `pause`, `resume`.
   - `audio/procedural.ts` — `tone`, `noise`, and their shared
     private envelope / output helpers.
   - `audio/audio.ts` — barrel of public re-exports plus the
     remaining lifecycle / track / mix / unload helpers.

   Public API surface (`me.audio.*`) is byte-for-byte identical.
   Cross-module mutable state lives in `state` (an object), avoiding
   the "ESM `let` exports don't share writes" footgun.

2. **`position()` / `orientation()` honest overloads.** Both functions
   previously claimed `: number[]` and `return … as unknown as number[]`,
   but Howler's `pos(x, y, z, id)` overload returns `this` (a Howl
   instance) — the cast was laundering a Howl as a number[]. Now they
   mirror Howler's actual API: overloaded as either `(name)` →
   `[number, number, number]` (group getter) or
   `(name, x, y?, z?, id?)` → `void` (setter). No more dishonest casts.

Build clean. Audio test count up to 36 (+5 for position/orientation:
existence + throw-on-missing-clip for both getter and setter shapes;
the round-trip path requires a loaded sound which the spec doesn't
set up, same as for every other file-playback function in the spec).
